feat: Generate asset generator index
