+++ server/routes/generate-assets/index.tsx
